refactoring: moved session container to session_container.rs.
